Revert "Add Maven To Project"
