+++ src/gasStationSoftware/controller/WindowController.java
@@ -1,10 +1,7 @@
 package gasStationSoftware.controller;
 
-import com.jfoenix.controls.JFXButton;
-import com.jfoenix.controls.JFXComboBox;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
@@ -21,15 +18,6 @@ public class WindowController implements Initializable {
 
     @FXML
     private AnchorPane userPane, sellingPane, inventoryPane, fuelTankPane, employeePane, reportPane, settingsPane;
-
-    @FXML
-    private JFXButton btnSettingsImport, btnSettingsExport, btnSettingsGasPumps, btnSettingsTanks, btnSettingsFuels, btnSettingsCreateTheme, btnSettingsEditTheme, btnSellingCheckOut, btnSellingGoods, btnSellingGasPumps;
-
-    @FXML
-    private JFXComboBox cbSettingsTheme, cbSettingsType;
-
-    @FXML
-    private Label lblUserName, lblUserRole, lblUserSumDay, lblUserSumMonth, lblUserSumYear, lblUserSumDayCurrency, lblUserSumMonthCurrency, lblUserSumYearCurrency, lblSellingPrice, lblSellingPriceCurrency;
 
     @FXML
     private void handleMenuButtonAction(MouseEvent event) {
