+++ src/gasStationSoftware/exceptions/NumberOutOfRangeException.java
@@ -1,0 +1,13 @@
+package gasStationSoftware.exceptions;
+
+public class NumberOutOfRangeException
+extends Exception {
+
+	public NumberOutOfRangeException() {
+		super();
+	}
+
+	public NumberOutOfRangeException(String message) {
+		super(message);
+	}
+}
