+++ src/gasStationSoftware/ui/resource/images/icons8-male-user-80.png
