+++ src/gasStationSoftware/ui/resource/images/icons8-oil-tanker-70.png
