+++ src/gasStationSoftware/ui/resource/images/icons8-people-70.png
