+++ src/gasStationSoftware/ui/resource/images/icons8-settings-70.png
