+++ src/gasStationSoftware/ui/resource/images/icons8-shopping-70.png
