+++ src/gasStationSoftware/ui/resource/images/icons8-system-report-70.png
