+++ src/gasStationSoftware/ui/resource/images/icons8-warehouse-70.png
