+++ src/gasStationSoftware/ui/resource/images/round-account-button-with-user-inside.png
