--- src/main/java/gasStationSoftware/exceptions/OSException.java
@@ -1,13 +1,0 @@
-package gasStationSoftware.exceptions;
-
-public class OSException
-        extends Exception {
-
-    public OSException() {
-        super();
-    }
-
-    public OSException(String message) {
-        super(message);
-    }
-}
